feat: add comprehensive test infrastructure and improve code quality
Build System
------------
- Add root Makefile with multi-device build support and unified test target
- Add per-device test Makefiles with Unity framework integration
- Add coverage reporting infrastructure (scripts/generate_coverage_report.py,
  scripts/aggregate_coverage_reports.py)
- Standardize Unity test framework paths and environment variable handling

New Features
------------
- Add ESM (Error Signal Monitor) module for LP8772x-Q1 with full API support
  for configuration, status monitoring, and error handling
- Add pmic_diag module for TPS65386x-Q1, relocating diagnostic functionality
  from pmic_core
- Implement architectural requirements across all devices with standardized
  diagnostic APIs (Pmic_getCommonStatus, Pmic_getCommonConfig, etc.)

Code Quality Improvements
-------------------------
- Fix MISRA.SHIFT.RANGE violations (248 issues) in header files by adding
  explicit casts to shift operations
- Fix Klocwork type safety violations (43 issues) with proper type handling
- Prevent NULL pointer dereferences across all device drivers
- Consolidate error codes into pmic_common.h, removing redundant pmic_status.h
- Modernize include guards to standard naming convention
- Standardize API const-correctness for configuration parameters
- Fix PMIC_STATUS macro to use direct masking instead of negation
- Add design and architecture tag linkages to all APIs

Documentation
-------------
- Update READMEs for new callback signatures and retry mechanism
- Add documentation for recently added modules
- Update examples with new parameters and signatures

Test Infrastructure
-------------------
- Add comprehensive test suites for all four PMIC devices:
  - LP8772x-Q1: Core, FSM, IO, IRQ, Power, WDG, ESM, Common modules
  - TPS65036x-Q1: Core, ESM, GPIO, IO, IRQ, Power, WDG, Common modules
  - TPS6522x-Q1: ADC, Core, ESM, FSM, GPIO, IO, IRQ, Power, WDG, Common modules
  - TPS65386x-Q1: Core, ESM, FSM, GPIO, IO, IRQ, Power, Timer, WDG, Common modules
- Add platform abstraction layer for mock I/O operations
- Add initialization and error path test coverage
- Add diagnostic API test coverage for pmic_common functions
- Add LCOV exclusions for unreachable defensive code validated by
  Pmic_checkHandle() in pmic_io.c and pmic_irq.c
- Enhance coverage script to support LCOV_EXCL_START/STOP blocks

+++ devices/LP8772x-Q1/include/regmap/core.h
@@ -40,12 +40,12 @@ extern "C" {
 #endif
 
 /* PMIC Module Device Revision Infos */
-#define PMIC_DEV_REV_REG        (0x01U)
+#define PMIC_DEV_REV_REG        ((uint8_t)0x01U)
 #define PMIC_DEV_REV_SHIFT      (0U)
 #define PMIC_DEV_REV_MASK       ((uint8_t)(0xFFU << PMIC_DEV_REV_SHIFT))
 
 /* PMIC Silicon Revision Info */
-#define PMIC_MANUFACTURING_VER_REG (0x04U)
+#define PMIC_MANUFACTURING_VER_REG ((uint8_t)0x04U)
 #define PMIC_SILICON_REV_SHIFT     (0U)
 #define PMIC_SILICON_REV_MASK      ((uint8_t)(0xFFU << PMIC_SILICON_REV_SHIFT))
 
@@ -73,7 +73,7 @@ extern "C" {
 #define CONFIG_CRC_STATUS_SHIFT         (2U)
 
 /* STAT_MODERATE_ERR definitions */
-#define STAT_MODERATE_ERR_REG           (0x59U)
+#define STAT_MODERATE_ERR_REG           ((uint8_t)0x59U)
 #define TSD_ORD_STAT_SHIFT              (0U)
 #define EXT_CLK_HIS_SHIFT               (1U)
 #define TRIM_TEST_CRC_STAT_SHIFT        (2U)
